changed prices types from decimal to integer

+++ src/main/java/raze/spring/inventory/domain/PurchaseTransaction.java
@@ -9,6 +9,7 @@ import javax.persistence.*;
 import javax.validation.constraints.Min;
 import java.math.BigDecimal;
 
+import java.math.BigInteger;
 import java.util.UUID;
 
 @Entity
@@ -25,7 +26,7 @@ public class PurchaseTransaction extends BaseEntity {
     private Long quantity;
 
     @Min(value = 0 , message = "must be greater than or equal 0")
-    private BigDecimal price;
+    private BigInteger price;
